+++ api/src/main/java/com/spring/winter/market/api/dtos/ProductDto.java
@@ -1,10 +1,12 @@
 package com.spring.winter.market.api.dtos;
 
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
     private String categoryTitle;
 
     public Long getId() {
@@ -23,11 +25,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
