+++ core_service/src/main/resources/db/migration/V1__init.sql
@@ -12,22 +12,22 @@ create table products
     id bigserial primary key,
     title varchar(255),
     category_id bigint references categories (id),
-    price int,
+    price double precision,
     created_at timestamp default current_timestamp,
     updated_at timestamp default current_timestamp
 );
 
 insert into products (title, price, category_id)
     values
-        ('milk', 80, 1),
-        ('Bread', 25, 1),
-        ('Cheese', 300, 1);
+        ('milk', 80.0, 1),
+        ('Bread', 25.0, 1),
+        ('Cheese', 300.0, 1);
 
 create table orders
 (
     id          bigserial primary key,
     username     varchar(255) not null,
-    total_price int not null,
+    total_price double precision not null,
     address      varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -40,8 +40,8 @@ create table order_items
     product_id          bigint not null references products (id),
     order_id            bigint not null references orders (id),
     quantity            int not null,
-    price_per_product   int not null,
-    price               int not null,
+    price_per_product   double precision not null,
+    price               double precision not null,
     created_at          timestamp default current_timestamp,
     updated_at          timestamp default current_timestamp
 );
